refactor(playback): ringbuf after decode: move messages to own file
